compose/exercise1/app/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
redis==5.0.3